+++ spec/spec_helper.rb
@@ -3,6 +3,7 @@
 ENV["RACK_ENV"] ||= "test"
 
 require "axn"
+require "axn/testing/spec_helpers"
 require "pry-byebug"
 
 $LOAD_PATH.unshift(File.expand_path(__dir__))
@@ -25,14 +26,6 @@ RSpec.configure do |config|
     end
   end
 end
-
-def build_action(&block)
-  action = Class.new.send(:include, Action)
-  action.class_eval(&block) if block
-  action
-end
-
-def build_axn(**kwargs, &) = Axn::Factory.build(**kwargs, &) # rubocop:disable Style/ArgumentsForwarding << not sure which ruby version that came in
 
 def build_interactor(*modules, &block)
   interactor = Class.new.send(:include, Interactor)